\name{gem_metadata}
\docType{data}
\alias{gem_metadata}
\title{Example Metadata from Gem Logger}
\description{
  Metadata (temperature, battery, etc.) from a Gem data logger. Provided
  as an example for running PlotMetadata.
}
\usage{gem_metadata}
\format{List containing vectors (time series) of different metadata types.}
\source{Recording made by author.}
\seealso{PlotMetadata}
\keyword{datasets}